# index header: Remove memWriter from fileWriter