<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

https://github.com/thanos-io/thanos/pull/5773 introduced a stateless mode to run store gateway and it embeds a new `MemWriter` into `FileWriter`.

Even in file writer mode, every time when writing, we need to write to 2 places, one is `MemWriter` and another one is `FileWriter`. This is unnecessary in non-stateless mode and caused us OOM kill when store gateway does initial sync or simply when resyncing blocks.

This pr simply removes `memWriter` from `fileWriter` because it is not used at all, but buffering more memory causing OOM kill.

Heap profile

![image](https://github.com/thanos-io/thanos/assets/25150124/45a164c2-5328-4327-a2a6-7d9110d58b2b)

## Verification

Existing tests should still pass.