@yeya24 2023-07-07
Benchmark 

```
name            old time/op    new time/op    delta
BinaryWrite-10    16.5ms ± 0%    17.3ms ± 0%   ~     (p=1.000 n=1+1)

name            old alloc/op   new alloc/op   delta
BinaryWrite-10    60.1kB ± 0%    72.0kB ± 0%   ~     (p=1.000 n=1+1)

name            old allocs/op  new allocs/op  delta
BinaryWrite-10      72.0 ± 0%      68.0 ± 0%   ~     (p=1.000 n=1+1)
```

Note that we have less allocs but more allocated bytes per op. I think the reason is that we are allocating a buf of 32KB when creating a file writer https://github.com/thanos-io/thanos/pull/6509/files#diff-57dac449d02a313b066d281b872c08ab719d9463345aea8a964a43c8e96f1e52R383.

In https://github.com/thanos-io/thanos/pull/5773, file writer uses the mem writer buffer. But mem writer buffer is not initialized so the size is 0. I think that's why this pr has more allocated bytes. And this behavior should be the same as the version before https://github.com/thanos-io/thanos/pull/5773.

Profile before this pr, able to see `MemWriter` allocating heap when writing. https://pprof.me/dac97d5/?filter_by_function=Mem&search_string=github.com%2Fthanos-io%2Fthanos%2Fpkg%2Fblock%2Findexheader.%28*MemoryWriter%29.Write&dashboard_items=icicle&dashboard_items=table

<img width="1830" alt="image" src="https://github.com/thanos-io/thanos/assets/25150124/c4d6508b-3ad1-4b33-944b-a2e7043a0b85">

With this pr, no `MemWriter` memory usage in heap anymore. https://pprof.me/7a315b0/?filter_by_function=Mem&search_string=Mem&dashboard_items=icicle&dashboard_items=table

<img width="1827" alt="image" src="https://github.com/thanos-io/thanos/assets/25150124/9e15c204-43a4-4b6f-af8e-6b080e76e9ff">