@yeya24 2023-07-07
Benchmark with this pr and commit `431a6cf0e876b4ddfd7c03559b99a34eef0c0db0` which is the one before #5773.

```
name            old time/op    new time/op    delta
BinaryWrite-10    14.7ms ± 0%    16.9ms ± 0%   ~     (p=1.000 n=1+1)

name            old alloc/op   new alloc/op   delta
BinaryWrite-10    71.9kB ± 0%    72.0kB ± 0%   ~     (p=1.000 n=1+1)

name            old allocs/op  new allocs/op  delta
BinaryWrite-10      63.0 ± 0%      68.0 ± 0%   ~     (p=1.000 n=1+1)
```

Previous version is faster and fewer allocs I guess because its logic is simpler. But alloc/op is similar to this one.